+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add overload of `DocusaurusMarkdownWriter.WriteDocusaurusFrontMatter` ([#38](https://github.com/josefpihrt/dotmarkdown/pull/38)).
+
 ## [0.3.0-beta] - 2023-07-27
 
 ### Added